Remove Django from dev requirements
It was impossible to run tox because it would complain about double
requirements for Django. Both in requirements_dev.txt and in the tox.ini
Django was given as requirement.

As tox is the documented approach to running the tests, I think it's
important to get it working even if it means one needs to install an
extra dependency if one wants to run the tests outside of tox.